requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
numpy==1.25.1
matplotlib==1.21.3
optuna==3.1.0
pandas==2.2.2
pillow==2.7.0
pyyaml==5.4.1
python-dotenv>=0.5.1
pytorch_lightning==1.7.0
scikit-learn==1.0.2
scipy==1.13.1
timm==1.0.9
torch==1.13.1
torchvision==0.14.1
torchmetrics==1.4.2